feat: FavoriteService 기능 추가

+++ src/main/java/nextstep/subway/favorite/dto/FavoriteRequest.java
@@ -1,22 +1,22 @@
 package nextstep.subway.favorite.dto;
 
 public class FavoriteRequest {
-    private Integer source;
-    private Integer target;
+    private Long source;
+    private Long target;
 
     public FavoriteRequest() {
     }
 
-    public FavoriteRequest(Integer source, Integer target) {
+    public FavoriteRequest(Long source, Long target) {
         this.source = source;
         this.target = target;
     }
 
-    public Integer getSource() {
+    public Long getSource() {
         return source;
     }
 
-    public Integer getTarget() {
+    public Long getTarget() {
         return target;
     }
 }
